chore: adjusts recursive function
